Improved coordinates (#2)
* Use BigDecimal for coordinates
* Save and load coordinates in SVG file with no precision loss
* Bump to Scala 3.7.1

+++ src/main/scala/io/github/scala_tessella/editor/interactions/TouchEventHandler.scala
@@ -114,7 +114,7 @@ object TouchEventHandler:
               scale = newScale,
               panX = newPanX,
               panY = newPanY
-            ).withRotation(newRotation))
+            ).withRotation(newRotation.toInt))
           }
         case _ => // State wasn't correctly initialized
